+++ src/types.ts
@@ -2,6 +2,10 @@ import { useActions } from './hooks/useActions';
 
 export type DrawIoEmbedProps = {
   /**
+   * This will cause the `onAutoSave` event to be triggered every time the diagram is changed
+   */
+  autosave?: boolean;
+  /**
    * Base URL for draw.io embed URL. Defaults to https://embed.diagrams.net
    */
   baseUrl?: string;
@@ -23,6 +27,7 @@ export type DrawIoEmbedProps = {
   configuration?: { [key: string]: any };
   exportFormat?: ExportFormats;
   onLoad?: (data: EventLoad) => void;
+  onAutoSave?: (data: EventAutoSave) => void;
   onSave?: (data: EventSave) => void;
   onClose?: (data: EventExit) => void;
   onConfigure?: (data: EventConfigure) => void;
@@ -35,6 +40,9 @@ export type DrawIoEmbedProps = {
 
 export type DrawIoEmbedRef = ReturnType<typeof useActions>;
 
+/**
+ * Part of the URL parameters based on https://www.drawio.com/doc/faq/supported-url-parameters
+ */
 export type UrlParameters = {
   /**
    * Theme of the editor
@@ -86,6 +94,59 @@ export type UrlParameters = {
    * @default false
    */
   noExitBtn?: boolean;
+  /**
+   * Uses the lightbox in chromeless mode (larger zoom, no page visible, chromeless
+   * 
+   * @default false
+   */
+  lightbox?: boolean;
+  /**
+   * Uses the chromeless read-only viewer.
+   * 
+   * @default false
+   */
+  chrome?: boolean;
+  /**
+   * Opens links in the same window or frame or in a blank window in chromeless mode (auto is the default which opens relative links and anchors in the same window in chromeless mode, and all links in a new window in editing mode)
+   * 
+   * @options auto|self|frame|blank
+   * @default auto
+   */
+  target?: string;
+  /**
+   * Adds a link for the Edit button in chromeless mode (use `edit=_blank` to edit the diagram as a new copy).
+   */
+  edit?: string;
+  /**
+   * Sets the default grid enabled to true.
+   * 
+   * @default false
+   */
+  grid?: boolean;
+  /**
+   * Enables folding in chromeless mode.
+   * 
+   * @default false
+   */
+  nav?: boolean;
+  /**
+   * Adds layer control in chromeless mode
+   * 
+   * @default false
+   */
+  layers?: boolean;
+  /**
+   * Space-separated list of layer IDs to show. If not specified, all layers are shown.
+   */
+  "layer-ids"?: string;
+  /**
+   * Shows the Close button in chromeless mode which closes the window when selected.
+   */
+  close?: boolean;
+  /**
+   * Specifies the language of the user interface. For possible values, see https://www.drawio.com/doc/faq/supported-url-parameters
+   */
+  lang?: string;
 };
 
 type ExportFormats = 'html' | 'html2' | 'svg' | 'xmlsvg' | 'png' | 'xmlpng';
@@ -93,6 +154,7 @@ type ExportFormats = 'html' | 'html2' | 'svg' | 'xmlsvg' | 'png' | 'xmlpng';
 export type EmbedEvents =
   | EventInit
   | EventLoad
+  | EventAutoSave
   | EventSave
   | EventExit
   | EventConfigure
@@ -118,6 +180,17 @@ export type EventSave = {
   xml: string;
   /** Is set when the event was triggered by anything other than the save action */
   parentEvent?: string;
+};
+
+export type EventAutoSave = {
+  event: 'autosave';
+  bounds: PagePosition;
+  currentPage: number;
+  page: PagePosition;
+  pageVisible: boolean;
+  scale: number;
+  translate: { x: number; y: number };
+  xml: string;
 };
 
 export type EventExit = {
@@ -182,6 +255,7 @@ export type ActionLoad = {
   xml?: string;
   xmlpng?: string;
   descriptor?: { format: 'csv'; data: string };
+  autosave?: boolean;
 };
 
 export type ActionMerge = {
@@ -272,4 +346,11 @@ export type ActionExport = {
   transparent?: boolean;
   /** Specifies the background color */
   background?: string;
+};
+
+type PagePosition = {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
 };